requirements.txt
autopep8==1.4.4
Babel==2.7.0
Django==2.2.6
django-phonenumber-field==3.0.1
djangorestframework==3.10.3
gunicorn==19.9.0
pep8==1.7.1
phonenumberslite==8.10.21
psycopg2==2.8.3
pycodestyle==2.5.0
pytz==2019.3
sqlparse==0.3.0
